+++ src/common/common.module.ts
@@ -1,9 +1,9 @@
 import { Global, Module } from '@nestjs/common';
-import ErrorHandlerService from './error.handler.service';
+import ErrorHandlerService from './utils/error.handler.service';
 
 @Global()
 @Module({
   providers: [ErrorHandlerService],
   exports: [ErrorHandlerService],
 })
-export class ErrorHandlerModule {}
+export class CommonModule {}
